+++ components/TaskEditModal.tsx
@@ -9,7 +9,7 @@ import { Textarea } from "@/components/ui/textarea"
 import { Task } from "@/lib/tasks"
 import { Tag } from "@/lib/tags"
 import { Eye } from "lucide-react"
-import { useToast } from "@/components/ui/use-toast"
+import { useToast } from "@/hooks/use-toast"
 import { cn } from "@/lib/utils"
 import { format } from "date-fns"
 import { Command, CommandEmpty, CommandGroup, CommandInput, CommandItem } from "@/components/ui/command"
